moved test to same folder as component
