+++ components/follow-button/follow-button.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FollowButton from '../follow-button';
+import FollowButton from './follow-button';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
